+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 1.3.46
+
+- `Json.defaultFieldValueResolver`:
+  - Improve resolution of `EntityReference` and ``.
+- reflection_factory: ^2.0.1
+
+## 1.3.45
+
+- `KeyCondition`: added support to `>`, `>=`, `<` and `<=` operators. 
+  - `KeyConditionGreaterThan`, `KeyConditionGreaterThanOrEqual`.
+  - `KeyConditionLessThan`, `KeyConditionLessThanOrEqual`.
+- `EntityHandler.createFromMap`: added parameter `jsonDecoder`.
+
 ## 1.3.44
 
 - `APIRouteRule`: